Fix for CustomObjectToImageConverter not working on all instances.
Fixed by moving it down a converter-layer to the base Convert function.
Reduced complexity of CustomObjectImageConverter to only take a value, and renamed it to make it more obvious that it might not return for a type.

Added new testdatat to both normal ribbon and commandbar which uses a CustomImage (which is just a most fancy textbox).

v7.0.0-0.62